+++ src/components/layout/Header.js
@@ -66,12 +66,6 @@ function Header() {
         }
     };
 
-    const onSaveMovements = async () => {
-        if (appApi.movementFile) {
-            await movementApi.saveMovementsToFile(appApi.movementFile, movementApi.movements);
-        }
-    };
-
     const onComputeCategories = async () => {
         await movementApi.computeCategories();
     };
@@ -81,7 +75,6 @@ function Header() {
 
         rule = await ruleApi.addRule(rule);
         ruleApi.setSelectedRuleIds(rule.id);
-        appApi.setEditingCell(rule.id, 'title');
     };
 
     const onOpenWebsite = async () => {
@@ -139,11 +132,10 @@ function Header() {
                 {createButton('cog', 'Préférences', onSetSettingsVisible, false)}
             </Button.Group>
             <Button.Group style={{ marginRight: 20 }}>
-                {appApi.selectedView === 'expenses' ? createButton('folder-open', 'Ouvrir le site', onOpenWebsite, false) : null}
-                {appApi.selectedView === 'expenses' ? createButton('pencil', 'Remplissage automatique', onAutoFill, false) : null}
+                {['expenses', 'income'].includes(appApi.selectedView) ? createButton('folder-open', 'Ouvrir le site', onOpenWebsite, false) : null}
+                {['expenses', 'income'].includes(appApi.selectedView) ? createButton('pencil', 'Remplissage automatique', onAutoFill, false) : null}
                 {appApi.selectedView === 'rules' ? createButton('plus', 'Ajouter une règle', onAddRule, false) : null}
                 {appApi.selectedView === 'movements' ? createButton('folder-open', 'Charger des mouvements', onLoadMovements, false) : null}
-                {appApi.selectedView === 'movements' ? createButton('save', 'Sauver les mouvements', onSaveMovements, false) : null}
                 {appApi.selectedView === 'movements' ? createButton('wand-magic-sparkles', 'Calculer les catégories', onComputeCategories, false) : null}
             </Button.Group>
         </>