+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# 1.1.0
+
+## New Features
+
+* Improve rule creation
+* Display rule as string
+* Automatically add amount condition
+* Save movement table scroll position
+
 # 1.0.0
 
 ## New Features